Add make-dir-complete plugin
This plugin is a dynamic directory completion generator. Given an alias
name and a directory, it creates a complete-enabled alias for the
chosen directory.

Example:

    make_dir_complete src ~/src/

This will create a command `src` that completes with the subdirectories
inside ~/src/. So, if you have something like:

    `-- src/
        |-- oh-my-zsh/
        |   `-- ...
        |-- oh-my-vim/
        |   `-- ...
        `-- whatever/
            `-- ...

Typing

    src oh<TAB>

Will give you a complete menu with "oh-my-zsh" and "oh-my-vim" as
options.

Ref: http://caioromao.com/blog/custom-directory-completion-for-zsh/